+++ CHANGELOG.md
@@ -4,8 +4,16 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
- 
+
 ## [Unreleased] - yyyy-mm-dd
+ 
+### Added
+
+### Changed
+
+### Fixed
+
+## [2.1.0] - 2025-07-28
  
 ### Added